Title: 최솟값 3개, Time: 670ms, Memory: 23MB, Status: Passed - CodeTree

+++ 241226/최솟값 3개/top-3-smallest-number.java
@@ -4,7 +4,7 @@ import java.util.PriorityQueue;
 
 public class Main {
     static int n;
-    static PriorityQueue<Integer> pq = new PriorityQueue<>();
+    static PriorityQueue<Long> pq = new PriorityQueue<>();
 
     public static void main(String[] args) throws Exception {
         StringTokenizer st;
@@ -14,15 +14,16 @@ public class Main {
 
         st = new StringTokenizer(br.readLine());
         for(int i=0; i<n; i++){
-            pq.offer(Integer.parseInt(st.nextToken()));
+            pq.offer(Long.parseLong(st.nextToken()));
 
             if (pq.size() <= 2){
                 System.out.println(-1);
                 continue;
             }
-            int a = pq.poll();
-            int b = pq.poll();
-            int c = pq.poll();
+            long a = pq.poll();
+            long b = pq.poll();
+            long c = pq.poll();
+
             System.out.println(a * b * c);
 
             pq.offer(a);
